initial commit for new context type and embedding of context within closures
This work introduces a redefined context type that associates a running frame
with methods for pausing, resuming and scheduling contexts, a fault handler,
and a transient heap that can be used for allocations that exist for the
lifetime of the context. kernel_contexts and unix threads now extend this base
type, and the syscall_context type has been created to encompass syscall
operations. Syscall operations which block awaiting continuation now stamp
their respective context in the continuation closure using
contextual_closure(). Areas of the kernel which service such continuations now
invoke them using async_apply*() functions, resulting in the context saved in
such closures to be resumed.

Now that each syscall in progress is represented by its own syscall_context,
it is possible to service concurrent page faults for multiple
syscalls. unix_fault_handler now handles faults in a way more orthogonal to
the context type, instead relying on the context methods to effectively
suspend and resume the context when blocking is necessary for faults on
file-backed memory.

Frame storage embedded within the context is now a fixed size, representing
only a basic frame save. The extended frame (FP / SIMD) storage is now
allocated separately, pointed to by the FRAME_EXTENDED field in the basic
frame.