Netlink read(): do not dequeue message if larger than read buffer
If a netlink socket contains an input message whose length is
larger than the available space in the user read buffer, the
message should not be dequeued from the socket queue, because it
cannot be read in its entirety.
This change fixes a bug in which a netlink message can be dequeued
(and the memory allocated for it can be leaked) without having been
read by userspace.