Sockets: add management of receiver buffer size
This commit changes the socket implementation (for network sockets,
Unix domain sockets, and netlink sockets) so that the total length
(in bytes) of the input data buffered in a given socket does not
exceed the value of the so_rcvbuf parameter (that is configurable
via a "so_rcvbuf" attribute in the manifest). The existing limit on
the maximum number of packets given by the size of the input queues
is kept in place.
In addition, any UDP packets that are discarded due the length of a
socket receive buffer being exceeded are now properly deallocated.
getsockopt() has been amended so that it returns the correct value
for the SO_SNDBUF and SO_RCVBUF socket options.
The network socket runtime test has been amended to exercise these
new features.